lets avoid nil

+++ db/migrate/20120412204330_recreate_gift_survey_questions.rb
@@ -1,8 +1,8 @@
 # -*- encoding : utf-8 -*-
 class RecreateGiftSurveyQuestions < ActiveRecord::Migration
   def up
-	  survey = Survey.find_by_name("Gift Survey")
-	  survey.questions.destroy_all
+    survey = Survey.find_by_name("Gift Survey")
+    survey.questions.destroy_all if survey
     SurveyBuilder.new(GIFT_SURVEY_DATA, "Gift Survey").build
   end
